# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateSubmissionRequest:
    properties:
      date:
        type: string
      field_id:
        type: string
      growth_stage:
        type: string
      location:
        type: string
      notes:
        type: string
      observer_name:
        type: string
      plant_conditions:
        items:
          type: string
        type: array
      trait_measurements:
        $ref: '#/definitions/models.TraitMeasurements'
    required:
    - date
    - field_id
    - growth_stage
    - location
    - observer_name
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  models.TraitMeasurements:
    properties:
      culm_length:
        type: number
      hills_observed:
        type: integer
      panicle_length:
        type: number
      panicles_per_hill:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a rice monitoring application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Rice Monitor API
  version: "1.0"
paths:
...
  /submissions:
    get:
      description: Get a list of all submissions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by submission status
        in: query
        name: status
        type: string
      - description: Filter by field ID
        in: query
        name: field_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all submissions
      tags:
      - submissions
    post:
      consumes:
      - application/json
      description: Create a new submission
      parameters:
      - description: Submission object that needs to be added
        in: body
        name: submission
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubmissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new submission
      tags:
      - submissions
  /submissions/{id}:
    delete:
      description: Delete a submission by its ID
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a submission
      tags:
      - submissions
    get:
      description: Get a single submission by its ID
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a submission by ID
      tags:
      - submissions
    put:
      consumes:
      - application/json
      description: Update an existing submission
      parameters:
      - description: Submission ID
        in: path
        name: id
        required: true
        type: string
      - description: Submission object that needs to be updated
        in: body
        name: submission
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a submission
      tags:
      - submissions
  /submissions/export:
    get:
      description: Export submissions to a CSV file
      produces:
      - text/csv
      responses:
        "200":
          description: CSV content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Export submissions to CSV
      tags:
      - submissions
...
schemes:
- http
- https
swagger: "2.0"